clean code

+++ main/src/loss/lovazs_loss.py
@@ -69,7 +69,7 @@ def iou_m(preds, labels, C, EMPTY=1., ignore=None, per_image=False):
                     iou.append(float(intersection) / union)
         ious.append(iou)
     ious = map(mean, zip(*ious)) # mean accross images if per_image
-    return 100 * np.fromiter(ious)
+    return 100 * np.fromiter(ious, dtype=np.float)
 
 
 # --------------------------- BINARY LOSSES ---------------------------
